Fixed an Issue in IOAddressSpace

+++ source/Cosmos/Cosmos.Hardware/PCIBus.cs
@@ -6778,7 +6778,7 @@ namespace Cosmos.Hardware {
                 else if ((address & PCI_BASE_ADDRESS_SPACE) == PCI_BASE_ADDRESS_SPACE_IO)
                 {
                     UInt32 size = (~(PCI_BASE_ADDRESS_IO_MASK & flags)) & 0xFF;
-                    IOMaps[i] = new Kernel.IOAddressSpace(address-1, ((uint)(byte)size));
+                    IOMaps[i] = new Kernel.IOAddressSpace(address-1, ((uint)size));
                     Console.Write("IOSpace IO. Size = ");
                     Interrupts.WriteNumber(size, 32);
                     Console.Write(", Address = ");
